Revert "Revert "Merge remote-tracking branch 'origin/develop' into develop""
This reverts commit 50fc775c69c123fd2b10c641685657413af58359.

+++ system/src/Grav/Common/Page/Medium/MediumFactory.php
@@ -119,8 +119,8 @@ class MediumFactory
         }
 
         $ratio = $to / $from;
-        $width = (int) ($medium->get('width') * $ratio);
-        $height = (int) ($medium->get('height') * $ratio);
+        $width = $medium->get('width') * $ratio;
+        $height = $medium->get('height') * $ratio;
 
         $prev_basename = $medium->get('basename');
         $basename = str_replace('@'.$from.'x', '@'.$to.'x', $prev_basename);
